Factor out sidebar buttons and move some hooks out of conditional statements
